fix some comments and print format

+++ rocketmq-spring-boot-samples/rocketmq-consume-demo/src/main/java/org/apache/rocketmq/samples/springboot/consumer/ObjectConsumerWithReplyUser.java
@@ -34,7 +34,7 @@ public class ObjectConsumerWithReplyUser implements RocketMQReplyListener<User, 
     public User onMessage(User user) {
         System.out.printf("------- ObjectConsumerWithReplyUser received: %s \n", user);
         User replyUser = new User();
-        replyUser.setUserAge(Byte.valueOf((byte) 10));
+        replyUser.setUserAge((byte) 10);
         replyUser.setUserName("replyUserName");
         return replyUser;
     }
